requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0